+++ src/components/Globe.tsx
@@ -14,12 +14,14 @@ type Props = {
   practiceMode: boolean;
 };
 
+const ZOOM_SPEED = 1;
+
 export default function Globe({ guesses, globeRef, practiceMode }: Props) {
   // State
   const [places, setPlaces] = useState(guesses);
 
   // Theme
-  const { nightMode, highContrast } = useContext(ThemeContext).theme;
+  const { nightMode, prideMode, highContrast } = useContext(ThemeContext).theme;
 
   // Globe size settings
   const size = isMobile ? 320 : 600; // px on one side
@@ -78,9 +80,21 @@ export default function Globe({ guesses, globeRef, practiceMode }: Props) {
       const answerCountry = JSON.parse(
         localStorage.getItem("practice") as string
       );
-      return getColour(country, answerCountry, nightMode, highContrast);
+      return getColour(
+        country,
+        answerCountry,
+        nightMode,
+        highContrast,
+        prideMode
+      );
     }
-    return getColour(country, answerCountry, nightMode, highContrast);
+    return getColour(
+      country,
+      answerCountry,
+      nightMode,
+      highContrast,
+      prideMode
+    );
   }
 
   // Label colour
@@ -111,6 +125,23 @@ export default function Globe({ guesses, globeRef, practiceMode }: Props) {
     const { altitude } = globeRef.current.pointOfView();
     coords["altitude"] = Math.max(altitude + z, 0.05);
     globeRef.current.pointOfView(coords, 250);
+  }
+
+  // Called when the globe position changes
+  function globeOnZoom() {
+    overrideGlobeZooming();
+  }
+
+  // Override the zoomSpeed mutation in globe.gl by calling this in the globe's
+  // onZoom callback.
+  //
+  // By the time this callback is called, an onchange event handler on
+  // `controls` defined in globe.gl's `globe.js` source file will have changed
+  // the zoomSpeed based on altitude. We will counteract that to get back the
+  // nice zooming implemented in the three.js library (`OrbitControls.js`).
+  function overrideGlobeZooming() {
+    const controls: any = globeRef.current?.controls();
+    if (controls != null) controls.zoomSpeed = ZOOM_SPEED;
   }
 
   const btnFill = nightMode ? "bg-[#582679]" : "bg-[#F3BC63]";
@@ -144,6 +175,7 @@ export default function Globe({ guesses, globeRef, practiceMode }: Props) {
           onPolygonClick={(p, e, c) => turnGlobe(c, globeRef)}
           polygonStrokeColor="#00000000"
           atmosphereColor={nightMode ? "rgba(63, 201, 255)" : "lightskyblue"}
+          onZoom={globeOnZoom}
         />
       </div>
       {isMobile && (