+++ src/components/List.tsx
@@ -2,10 +2,11 @@ import { SyntheticEvent, useContext, useEffect, useState } from "react";
 import { GlobeMethods } from "react-globe.gl";
 import { Country, LanguageName } from "../lib/country";
 import { answerName } from "../util/answer";
-import { findCentre, turnGlobe, altitudeFunction } from "../util/globe";
+import { findCentre, turnGlobe } from "../util/globe";
 import { LocaleContext } from "../i18n/LocaleContext";
 import { Locale } from "../lib/locale";
 import { FormattedMessage } from "react-intl";
+import Toggle from "./Toggle";
 
 type Props = {
   guesses: Country[];
@@ -27,6 +28,7 @@ function reorderGuesses(guessList: Country[]) {
 
 export default function List({ guesses, win, globeRef }: Props) {
   const [orderedGuesses, setOrderedGuesses] = useState(reorderGuesses(guesses));
+  const [miles, setMiles] = useState(false);
   const { locale } = useContext(LocaleContext);
   const langNameMap: Record<Locale, LanguageName> = {
     "pt-BR": "NAME_PT",
@@ -34,6 +36,8 @@ export default function List({ guesses, win, globeRef }: Props) {
     "en-CA": "NAME_EN",
     "fr-FR": "NAME_FR",
     "de-DE": "NAME_DE",
+    "pl-PL": "NAME_PL",
+    "it-IT": "NAME_IT",
   };
   const langName = langNameMap[locale];
 
@@ -41,9 +45,18 @@ export default function List({ guesses, win, globeRef }: Props) {
     setOrderedGuesses(reorderGuesses(guesses));
   }, [guesses]);
 
-  function formatKm(m: number) {
-    const km = m / 1000;
-    return km.toFixed(0).replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+  function formatKm(m: number, miles: boolean) {
+    const METERS_PER_MILE = 1609.34;
+    const BIN = 10;
+    const value = miles ? m / METERS_PER_MILE : m / 1000;
+    if (value < BIN) return "< " + BIN;
+
+    const rounded = Math.round(value / BIN) * BIN;
+    // const max = min + BIN;
+    const format = (num: number) =>
+      num.toFixed(0).replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+
+    return `~ ${format(rounded)}`;
   }
 
   const qualifier = win ? "Answer" : "Closest";
@@ -57,15 +70,24 @@ export default function List({ guesses, win, globeRef }: Props) {
   const closest = orderedGuesses[0];
   const farthest = orderedGuesses[orderedGuesses.length - 1];
 
+  const [isSortedByDistance, setIsSortedByDistance] = useState(true);
+  const guessesToDisplay = isSortedByDistance ? orderedGuesses : guesses;
+
   return (
     <div className="md:ml-10 md:mr-0 py-8 dark:text-white z-30">
       {orderedGuesses.length > 0 && (
         <p className="my-1">
-          <b>{qualifier}</b>
+          {isSortedByDistance ? (
+            <b>
+              <FormattedMessage id={qualifier} />
+            </b>
+          ) : (
+            <b>Guessed</b>
+          )}
         </p>
       )}
       <ul className="grid grid-cols-3 md:grid-cols-4 gap-3">
-        {orderedGuesses.map((guess, idx) => {
+        {guessesToDisplay.map((guess, idx) => {
           const { NAME_LEN, ABBREV, NAME, FLAG } = guess.properties;
           const flag = (FLAG || "").toLocaleLowerCase();
           let name = NAME_LEN >= 10 ? ABBREV : NAME;
@@ -84,7 +106,7 @@ export default function List({ guesses, win, globeRef }: Props) {
                   alt={name}
                   className=""
                 />
-                <span className="ml-2 text-md">{name}</span>
+                <span className="ml-2 text-md text-left">{name}</span>
               </button>
             </li>
           );
@@ -92,8 +114,30 @@ export default function List({ guesses, win, globeRef }: Props) {
       </ul>
       {closest && farthest && (
         <div className="mt-8">
+          <div className="flex items-center space-x-1">
+            <p>
+              <FormattedMessage id="Game8" />:{" "}
+              {formatKm(closest?.proximity, miles)}
+            </p>
+            <Toggle
+              name="miles"
+              setToggle={setMiles}
+              toggle={miles}
+              on="km"
+              off="miles"
+            />
+          </div>
           <p>
-            <FormattedMessage id="Game8" />: {formatKm(closest?.proximity)} km
+            <button
+              onClick={() => setIsSortedByDistance(!isSortedByDistance)}
+              className="mt-2"
+            >
+              <span className="text-md underline">
+                {isSortedByDistance
+                  ? "Sort by order of guesses"
+                  : "Sort by distance"}
+              </span>
+            </button>
           </p>
         </div>
       )}