+++ src/components/Outline.tsx
@@ -13,8 +13,7 @@ type Props = {
 };
 
 export default function Outline({ countryName, width }: Props) {
-  const { nightMode } = useContext(ThemeContext).theme;
-  const { highContrast } = useContext(ThemeContext).theme;
+  const { nightMode, highContrast, prideMode } = useContext(ThemeContext).theme;
 
   const country = countryData.find((p) => p.properties.NAME === countryName);
   if (!country)
@@ -32,7 +31,13 @@ export default function Outline({ countryName, width }: Props) {
 
   const outline = getPath(countryName);
 
-  const colour = getColour(countryCopy, sampleAnswer, nightMode, highContrast);
+  const colour = getColour(
+    countryCopy,
+    sampleAnswer,
+    nightMode,
+    highContrast,
+    prideMode
+  );
 
   return (
     <figure