+++ src/i18n/locales.ts
@@ -1,15 +1,11 @@
 import { LanguageName } from "../lib/country";
-import { Locale, LocaleMessages } from "../lib/locale";
+import { Locale } from "../lib/locale";
 import { English } from "./messages/en-CA";
 import { Spanish } from "./messages/es-MX";
 import { French } from "./messages/fr-FR";
 import { German } from "./messages/de-DE";
 import { Portuguese } from "./messages/pt-BR";
-
-// export const LOCALES: LocaleList = {
-//   English: "English",
-//   Spanish: "Spanish",
-// };
+import { Polish } from "./messages/pl-PL";
 
 export const LOCALES = {
   English: English,
@@ -17,6 +13,7 @@ export const LOCALES = {
   French: French,
   German: German,
   Portuguese: Portuguese,
+  Polish: Polish,
 };
 
 export const langNameMap: Record<Locale, LanguageName> = {
@@ -25,4 +22,6 @@ export const langNameMap: Record<Locale, LanguageName> = {
   "fr-FR": "NAME_FR",
   "de-DE": "NAME_DE",
   "pt-BR": "NAME_PT",
+  "pl-PL": "NAME_PL",
+  "it-IT": "NAME_IT",
 };