+++ src/i18n/messages/index.ts
@@ -4,12 +4,16 @@ import { Spanish } from "./es-MX";
 import { French } from "./fr-FR";
 import { German } from "./de-DE";
 import { Portuguese } from "./pt-BR";
+import { Italian } from "./it_IT";
+import { Polish } from "./pl-PL";
 
 const localeList: LocaleMessages = {
   "en-CA": English,
+  "de-DE": German,
   "es-MX": Spanish,
   "fr-FR": French,
-  "de-DE": German,
+  "it-IT": Italian,
+  "pl-PL": Polish,
   "pt-BR": Portuguese,
 };