+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change log
 
+## v1.6.0 - May 31, 2022
+- Added translations for Italian and Polish (thank you international translators!)
+- Added rainbow theme for pride 🏳️‍🌈
+- Increased zoom speed
+- You can now toggle the list of guessed countries between "sort by order of guesses" and "sort by distance"
+- Added switch for miles to km in closest border preview
+- Rounded closest border preview to nearest 10 km/miles due to imprecisions in the game's borders
+- Added (shameful self-promotion) link to [Plurality](https://plurality.fun) to the stats menu
+- Added fallback error message to share score functionality
+- Added alternate names for some French, Spanish, and Portuguese countries
+
 ## v1.5.0 - Apr 25, 2022
 - Added Practice mode so players can play unlimited times per day to hone their geography skills without affecting their score
 - Added translations for French, German, and Portuguese (thank you international translators!)